requirements.txt
beautifulsoup4==4.11.1
langchain==0.0.218
matplotlib==3.7.1
openai==0.27.8
pandas==1.5.3
pinecone_client==2.2.1
PyYAML==6.0
PyYAML==6.0
Requests==2.31.0
schedule==1.2.0
SQLAlchemy==2.0.16
streamlit
wordcloud==1.9.2
mysql-connector-python
mplfinance
akshare
pywencai
scikit-learn
websocket-client